modules/LocationEventConsumer/requirements.txt
SQLAlchemy==1.4.27
greenlet==1.1.2
kafka-python==2.0.2
psycopg2-binary==2.9.2
